requirements.txt
comtypes
faster-whisper
numpy
openai
pillow
pyautogui
pycaw
pynput
pyperclip
pyspellchecker
pystray
python-dotenv
pywin32
requests
sounddevice
soundfile
tomli-w
webrtcvad
